Refactor validation object to make it immutable
The immutable approach is a good thing for code clarity, functional
style and a lot of other improvements. But the most important benefit
of  this change it to make validation objects able to be `forked`
into another validation objects, like this:

const a = v8n()
  .number();

const b = a.min(40);

This kind of behavior was not possible in the previous mutable
implementation as the `b.min(40)` call was also changing the `a`
strategy.

In the new immutable approach, now `a` and `b` are different in their
behaviors:

a.test(40); // true
b.test(40); // false